Провел рефакторинг сервиса transfer

+++ transfer/src/main/java/com/bank/transfer/dto/AccountBalanceUpdateRequest.java
@@ -2,8 +2,10 @@ package com.bank.transfer.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountBalanceUpdateRequest(
         String currency,
-        double balance
+        BigDecimal balance
 ) {}
